Rename chanjo api
